Passing tests

+++ src/service/PerformanceService.ts
@@ -15,6 +15,7 @@ import Database from '../database/Database';
  **/
 export async function createComment(id: Number, comment: IComment, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
         const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL create_performance_review_comment(?,?,?,?)', [
             id,
@@ -40,6 +41,11 @@ export async function createComment(id: Number, comment: IComment, xAuthToken: s
  **/
 export async function deleteComment(id: Number, idComment: Number, xAuthToken: string) {
     try {
+        const comment_ = await getComment(id, idComment, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
+        if(comment_.fkCommenter != client.id && client.adminLevel != IEmployee.adminLevelEnum.HR_ADMIN){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'Can not edit a comment that is not yours.')
+        }
         let res = await Database.getInstance().query('CALL delete_performance_comment(?)', [
             idComment
         ], JABCResponse.PERFORMANCE)
@@ -60,6 +66,7 @@ export async function deleteComment(id: Number, idComment: Number, xAuthToken: s
  **/
 export async function deletePerformancePlan(id: Number, xAuthToken: string) {
     try {
+        const performancePlan = await getPerformancePlan(id, xAuthToken)
 		await Database.getInstance().query('CALL delete_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance plan was successfully deleted')
 	} catch (error) {
@@ -78,6 +85,7 @@ export async function deletePerformancePlan(id: Number, xAuthToken: string) {
  **/
 export async function deletePerformanceReview(id: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
 		await Database.getInstance().query('CALL delete_performance_review(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance review was successfully deleted')
 	} catch (error) {
@@ -97,8 +105,9 @@ export async function deletePerformanceReview(id: Number, xAuthToken: string) {
  **/
 export async function getComment(id: Number, idComment: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
         let res = await Database.getInstance().query('CALL get_comment(?)', [idComment], JABCResponse.PERFORMANCE)
-		return new Comment(res[0][0][0])
+        return new Comment(res[0][0][0])
     } catch (error) {
         throw error;
     }
@@ -115,6 +124,7 @@ export async function getComment(id: Number, idComment: Number, xAuthToken: stri
  **/
 export async function getComments(id: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
         let res = await Database.getInstance().query('CALL get_performance_review_comments(?)', [id], JABCResponse.PERFORMANCE)
 		return Comment.Comments(res[0][0])
     } catch (error) {
@@ -133,8 +143,24 @@ export async function getComments(id: Number, xAuthToken: string) {
  **/
 export async function getPerformancePlan(id: Number, xAuthToken: string) {
     try {
+        const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
         let performancePlan = new PerformancePlan(res[0][0][0])
+        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+			let managed = false;
+			let managersOfEmployee = await getManagersByEmployee(performancePlan.fkEmployee, xAuthToken)
+			for (let manager of managersOfEmployee) {
+				if (manager.id === client.id) {
+					managed = true;
+					break;
+				}
+			}
+			if (!managed) {
+				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
+			}
+		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performancePlan.fkEmployee){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance plan.')
+        }
         performancePlan.sections = PerformanceSection.PerformanceSections(res[0][1])
         performancePlan.comments = Comment.Comments(res[0][2])
         return performancePlan
@@ -154,8 +180,24 @@ export async function getPerformancePlan(id: Number, xAuthToken: string) {
  **/
 export async function getPerformanceReview(id: Number, xAuthToken: string) {
     try {
+        const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_review(?)', [id], JABCResponse.PERFORMANCE)
         let performanceReview = new PerformanceReview(res[0][0][0])
+        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+			let managed = false;
+			let managersOfEmployee = await getManagersByEmployee(performanceReview.fkEmployee, xAuthToken)
+			for (let manager of managersOfEmployee) {
+				if (manager.id === client.id) {
+					managed = true;
+					break;
+				}
+			}
+			if (!managed) {
+				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
+			}
+		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performanceReview.fkEmployee){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance review.')
+        }
         performanceReview.sections = PerformanceSection.PerformanceSections(res[0][1])
         performanceReview.comments = Comment.Comments(res[0][2])
         return performanceReview
@@ -177,11 +219,16 @@ export async function getPerformanceReview(id: Number, xAuthToken: string) {
  **/
 export async function updateComment(id: Number, idComment: Number, comment: IComment, xAuthToken: string) {
     try {
+        const comment_ = await getComment(id, idComment, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
+        if(comment_.fkCommenter != client.id && client.adminLevel != IEmployee.adminLevelEnum.HR_ADMIN){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'Can not edit a comment that is not yours.')
+        }
         let res = await Database.getInstance().query('CALL update_comment(?,?,?,?)', [
             idComment,
             comment.comment,
             comment.date,
-            comment.fkCommenter,
+            client.id
         ], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.PERFORMANCE, `The comment was edited successfully`)
     } catch (error) {
@@ -202,22 +249,7 @@ export async function updateComment(id: Number, idComment: Number, comment: ICom
 export async function updatePerformancePlan(id: Number, performancePlan: IPerformancePlan, xAuthToken: string) {
     let db;
 	try {
-		const client = (await Auth(xAuthToken)).employee
-		if(client.adminLevel == IEmployee.adminLevelEnum.MANAGER){
-			let managed = false;
-			let managersOfEmployee = await getManagersByEmployee(id, xAuthToken)
-			for(let manager of managersOfEmployee){
-				if(manager.id === client.id){
-					managed = true;
-					break;
-				}
-			}
-			if(!managed){
-				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
-			}
-		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id !== id){
-			throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF level, can not update a performance plan of other employee.')
-		}
+		const performancePlan_ = await getPerformancePlan(id, xAuthToken)
 		db = Database.getInstance();
         await db.beginTransaction();
 
@@ -229,12 +261,14 @@ export async function updatePerformancePlan(id: Number, performancePlan: IPerfor
 
 		await db.rawQuery('CALL delete_plan_sections(?)', [id], JABCResponse.PERFORMANCE);
 
-        for (let section of performancePlan.sections) {
-            await db.rawQuery('CALL create_employee_performance_plan_section(?,?,?)', [
-                id,
-                section.data,
-                section.sectionName
-            ], JABCResponse.PERFORMANCE);
+        if (performancePlan.sections != null){
+            for (let section of performancePlan.sections) {
+                await db.rawQuery('CALL create_employee_performance_plan_section(?,?,?)', [
+                    id,
+                    section.data,
+                    section.sectionName
+                ], JABCResponse.PERFORMANCE);
+            }
         }
 
         await db.commit();
@@ -262,22 +296,7 @@ export async function updatePerformancePlan(id: Number, performancePlan: IPerfor
 export async function updatePerformanceReview(id: Number, performanceReview: IPerformanceReview, xAuthToken: string) {
     let db;
 	try {
-		const client = (await Auth(xAuthToken)).employee
-		if(client.adminLevel == IEmployee.adminLevelEnum.MANAGER){
-			let managed = false;
-			let managersOfEmployee = await getManagersByEmployee(id, xAuthToken)
-			for(let manager of managersOfEmployee){
-				if(manager.id === client.id){
-					managed = true;
-					break;
-				}
-			}
-			if(!managed){
-				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
-			}
-		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id !== id){
-			throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF level, can not update a performance review of other employee.')
-		}
+		const performanceReview_ = await getPerformanceReview(id, xAuthToken)
 		db = Database.getInstance();
         await db.beginTransaction();
 
@@ -289,12 +308,14 @@ export async function updatePerformanceReview(id: Number, performanceReview: IPe
 
         await db.rawQuery('CALL delete_review_sections(?)', [id], JABCResponse.PERFORMANCE);
 
-        for (let section of performanceReview.sections) {
-            await db.rawQuery('CALL create_employee_performance_review_section(?,?,?)', [
-                id,
-                section.data,
-                section.sectionName
-            ], JABCResponse.PERFORMANCE);
+        if (performanceReview.sections != null){
+            for (let section of performanceReview.sections) {
+                await db.rawQuery('CALL create_employee_performance_review_section(?,?,?)', [
+                    id,
+                    section.data,
+                    section.sectionName
+                ], JABCResponse.PERFORMANCE);
+            }
         }
 
         await db.commit();
